Update: Fix compiler warning
git-svn-id: https://svn.code.sf.net/p/moneymanagerex/code/trunk@2545 7f5d68d0-c712-0410-839f-9b116a576cc5

+++ mmex/src/dbwrapper.cpp
@@ -404,7 +404,7 @@ void mmDBWrapper::createCurrencyV1Table(wxSQLite3Database* db)
         }
     }
 
-    for (int i = 0; i < currencies.Count(); ++i)
+    for (size_t i = 0; i < currencies.Count(); ++i)
     {
         wxStringTokenizer tk(currencies[i], wxT(";"));   
         DB_View_CURRENCYFORMATS_V1::Data* currency_data = CURRENCYFORMATS_V1.create();
